Fixes #11323 - fixed PID writing, interrupt trap and daemon logging
